fix bug

+++ app/controllers/appointments_controller.rb
@@ -45,7 +45,7 @@ class AppointmentsController < ApplicationController
     if user.present?
       today = Time.zone.now.strftime("%Y-%m-%d")
       appointment = Appointment.where(user_id: user.id, date: today).first
-      appointment.destroy
+      appointment.destroy if appointment.present?
     else
       render json: {}, status: :unauthorized
     end
